+++ CHANGELOG.md
@@ -11,7 +11,8 @@
 - Update AnnData loader to handle artbitrary gene slicing.
 - Do not do call `makeDefaultGetCellIsSelected` if not necessary.
 - Fix `categories` parsing for AnnData with `dtype=|O1`.
-
+- Use `concaveman` instead of `@turf/concave`.
+ 
 ## [1.1.6](https://www.npmjs.com/package/vitessce/v/1.1.6) - 2021-03-05
 
 ### Added